apps/finance-trading/requirements-test.txt
# Testing Dependencies
# Comprehensive testing suite for CI/CD pipeline

# =============================================================================
# CORE TESTING FRAMEWORK
# =============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
pytest-benchmark==4.0.0
pytest-timeout==2.2.0
pytest-env==1.1.3
pytest-html==4.1.1
pytest-json-report==1.5.0

# =============================================================================
# DATABASE TESTING
# =============================================================================
pytest-postgresql==5.0.0
testing.postgresql==1.3.0
alembic-postgresql-enum==1.1.2
sqlalchemy-utils==0.41.1

# =============================================================================
# HTTP TESTING
# =============================================================================
httpx==0.25.2
responses==0.24.1
aioresponses==0.7.4
pytest-httpx==0.28.0

# =============================================================================
# MOCK AND FIXTURES
# =============================================================================
factory-boy==3.3.0
faker==20.1.0
freezegun==1.2.2
pytest-factoryboy==2.5.1

# =============================================================================
# PROPERTY-BASED TESTING
# =============================================================================
hypothesis==6.92.1
hypothesis-pytest==0.19.0

# =============================================================================
# LOAD AND PERFORMANCE TESTING
# =============================================================================
locust==2.17.0
pytest-benchmark==4.0.0
pytest-profiling==1.7.0

# =============================================================================
# INTEGRATION TESTING
# =============================================================================
testcontainers==3.7.1
docker==6.1.3
docker-compose==1.29.2

# =============================================================================
# COMPLIANCE AND SECURITY TESTING
# =============================================================================
bandit==1.7.5
safety==2.3.5
pytest-security==0.1.0

# =============================================================================
# CODE COVERAGE
# =============================================================================
coverage==7.3.2
pytest-cov==4.1.0
coverage-badge==1.1.0

# =============================================================================
# SNAPSHOT TESTING
# =============================================================================
syrupy==4.6.0

# =============================================================================
# MARKET DATA TESTING
# =============================================================================
pandas-market-calendars==4.3.3
yfinance==0.2.28

# =============================================================================
# FINANCIAL TESTING UTILITIES
# =============================================================================
quantlib==1.32
numpy==1.26.2
pandas==2.1.4

# =============================================================================
# WEBSOCKET TESTING
# =============================================================================
websockets==12.0
pytest-websocket==0.1.0

# =============================================================================
# REDIS TESTING
# =============================================================================
fakeredis==2.20.1
redis==5.0.1

# =============================================================================
# COMPLIANCE TESTING
# =============================================================================
xmltodict==0.13.0
jsonschema==4.20.0
cerberus==1.3.5

# =============================================================================
# REPORT GENERATION
# =============================================================================
allure-pytest==2.13.2
pytest-html==4.1.1
junit-xml==1.9

# =============================================================================
# STRESS TESTING
# =============================================================================
pytest-stress==1.0.1
pytest-repeat==0.9.3

# =============================================================================
# PARALLEL TESTING
# =============================================================================
pytest-xdist==3.5.0
pytest-parallel==0.1.1

# =============================================================================
# TIME-BASED TESTING
# =============================================================================
time-machine==2.13.0
pytest-freezegun==0.4.2

# =============================================================================
# MONITORING TESTING
# =============================================================================
prometheus-client==0.19.0
opentelemetry-test-utils==0.42b0